requirements.txt
streamlit==1.25.0
pytesseract==0.3.10
spacy==3.7.3
ollama==0.1.3
Pillow==10.0.0

